Update Robolectric Activity tests
Robolectric.setupActivity() is deprecated and replaced with stepping
through the Activity lifecycle manually through ActivityController. The
preferred alternative of using ActivityScenario is not valid for these
cases as ActivityScenario places tighter restrictions around lifecycle
transitions, particularly that the Activity needs to be driven to a
'steady' state before it can be referenced or acted upon. This is
problematic since we immediately finish these Activities in onCreate().
Test has been updated to verify finished/finishing state.